codigo_projeto/requirements.txt
numpy==1.25.1
opencv_python==4.8.0.76
Pillow==10.0.0
